% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reanalysis.R
\name{AddReanalysis}
\alias{AddReanalysis}
\title{Extracts reanalysis output to match a set of station data.}
\usage{
AddReanalysis(obs, var = "air.2m", type = "mean", version = "3.5.1")
}
\arguments{
\item{obs}{- list of station data from \code{\link{ReadObs}}.}

\item{var}{- Which reanalysis variable (default 'air.2m').}

\item{type}{- What reanalysis data is required? 'mean' (default),
'spread', or 'normal'}

\item{version}{- Which version of 20CR (ensda run number), defaults
to '3.5.1' (v2c).}
}
\value{
list - as input, but for each station a new component
                'reanalysis.mean' (or '.normal' or '.spread'),
                a data frame with columns 'chron' and var.
}
\description{
Takes a station list from ReadObs, and extracts, for each station,
 reanalysis actuals and normals at 0,6,12 & 18 hours for the whole
 timespan of the obs data.
}
\details{
Currently only does air temperature.
}

