% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{FindStations}
\alias{FindStations}
\title{Find stations with data in a given region and date range.}
\usage{
FindStations(dir, lat.range = c(-90, 90), lon.range = c(-180, 180),
  date.range = c("1000-01-01:00", "3000-12-31:23"), meta = NULL)
}
\arguments{
\item{dir}{- directory containing ISD-lite}

\item{lat.range}{- Selected latitudes, defaults to c(-90.90).}

\item{lon.range}{- Selected longitudes, defaults to c(-180,180).}

\item{date.range}{- start and end dates in format
c(YYYY-MM-DD:HH,YYYY-MM-DD:HH) - defaults to all years.}

\item{meta}{- Metadata from \code{\link{ReadMetaData}} (calculated if NULL (default)).}
}
\value{
list - 'USAF', 'WBAN', 'LAT', 'LON', & 'ELEV' (subset of metadata).
}
\description{
Locates stations by lat.range, lon.range, and date.range.
}
\details{
Finds stations in the metadata file, not available on the local system.
}

