% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reanalysis.R
\name{GetVarAtHour}
\alias{GetVarAtHour}
\title{Get ISD data interpolated to a given hour.}
\usage{
GetVarAtHour(obs, year, month, day, hour, var = "AT", type = "mean")
}
\arguments{
\item{obs}{- list of station data from \code{\link{ReadObs}}.}

\item{var}{- Which variable: 'AT' (default),'DPT','SLP','WD',
'WS','TCC','LP1H','LP6H', 'air.2m', ...}

\item{type}{- 'actual' (default, and only option for observed values),
'spread', or 'normal'}
}
\value{
values - for each station in the input, selected value interpolated
                  to specified hour
}
\description{
First load the station obs with \code{\link{ReadObs}}, then add the
 reanalysis equivalents as required with calls to \code{\link{AddReanalysis}},
 then call this function to get interpolated values at any hour.
}
\details{
Currently only does air temperature.
}

