% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{ReadMetaData}
\alias{ReadMetaData}
\title{Read in the station metadata: Numbers, name, lat, lon, height for
all stations. (Reads the isd-history.txt file).}
\usage{
ReadMetaData(dir)
}
\arguments{
\item{dir}{- directory containing ISD-lite}
}
\value{
data frame - 1 row per station, columns as in the ISD-lite
 documentation.
}
\description{
There's a problem with some of the latitude's in the file - R
 does not think they are valid numbers. These are forced to NA
 but mean the function will throw some warnings.
}
\details{
Returns a data frame containing the metadata for each station
}

