% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{ReadObs}
\alias{ReadObs}
\title{Reads all observations in a given region and date range.}
\usage{
ReadObs(dir, lat.range, lon.range, date.range, meta = NULL)
}
\arguments{
\item{dir}{- directory containing ISD-lite}

\item{lat.range}{- Selected latitudes, defaults to c(-90.90).}

\item{lon.range}{- Selected longitudes, defaults to c(-180,180).}

\item{date.range}{- start and end dates in format
c(YYYY-MM-DD:HH,YYYY-MM-DD:HH)}

\item{meta}{- Metadata from \code{\link{ReadMetaData}} (calculated if NULL (default)).}
}
\value{
list - Each element 1 station list from \code{\link{ReadStation}}.
}
\description{
Locates stations by \code{\link{ReadMetaData}}, reads each one with \code{\link{ReadStation}}.
}
\details{
Finds stations in the metadata file, not available on the local system.
}

