% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{ReadStationFile}
\alias{ReadStationFile}
\title{Read in the hourly observations from one station file.}
\usage{
ReadStationFile(fname)
}
\arguments{
\item{fname}{- name of file to be read}
}
\value{
data frame - Observations: 1 row per hour, columns as in the ISD-lite
 documentation, except that AT, DPT, and WS are scaled to degC and m/s,
   and there is an added 'chron' column with the date-times.
}
\description{
Reads all the observations from one file (1 year, 1 station).
}
\details{
Most users will want to call ReadStation instead.
}

