% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{ReadStation}
\alias{ReadStation}
\title{Read in the hourly observations from one station for a given period.}
\usage{
ReadStation(usaf, wban, date.range = c("1000-01-01:00", "3000-12-31:23"),
  meta = NULL)
}
\arguments{
\item{usaf}{- USAF number of station.}

\item{wban}{- WBAN number of station.}

\item{date.range}{- start and end dates in format
c(YYYY-MM-DD:HH,YYYY-MM-DD:HH)}

\item{meta}{- Metadata from \code{\link{ReadMetaData}} (calculated if NULL (default)).}
}
\value{
list - 'data' - observations as from \code{\link{ReadStationFile}},
    also 'USAF', 'WBAN', 'LAT', 'LON', & 'ELEV'.
}
\description{
Identifies the station by combination of USAF and WBAN numbers
  (need both).
}
\details{
Combines data from more than one file, where necessary.
}

